Restructuring of classes and minor refactoring
